# Skip tessellation and merge vertices during CSG output.