… .github/workflows/build-docs.yml
… CHANGELOG.md
… Elements/src/Geometry/Mesh.cs
… Elements/src/Geometry/Solids/Solid.cs
… Elements/src/GraphicsBuffers.cs
… Elements/src/Model.cs
… Elements/src/StructuralFraming.cs
… Elements/test/CsgTests.cs
… Elements/test/MassTests.cs
… Elements/test/StructuralFramingTests.cs