BACKGROUND:
There are a number of places in our glTF export code where we unnecessarily tesselate CSG polygons. This is slow. We also don't merge vertices during export leading to a jagged appearance of most geometry.

DESCRIPTION:
This PR skips tessellating CSG polygons except when completely necessary, and uses the Octree introduced in a previous PR to enable fast merging of vertices. This results in near equivalent time CSG output and lower memory usage both during CSG writes and for the exported glTF.

For our test which generates beams with every HSS section and exports the results to glTF, the results are as follows.

|                  Method |     Mean |    Error |   StdDev |      Gen 0 |     Gen 1 |     Gen 2 | Allocated |.   File Size|
|------------------------ |---------:|---------:|---------:|-----------:|----------:|----------:|----------:|----:|
| 'Create all HSS beams.' | 306.5 ms | 34.32 ms | 22.70 ms | 55000.0000 | 5000.0000 | 2000.0000 | 205.99 MB |993kb|
| 'Create all HSS beams.' | **383.5 ms** | 12.87 ms | 8.51 ms | 41000.0000 | 2000.0000 |     - | **140.97 MB** |**667kb**|

TESTING:
- The table above is created by `Elements.Benchmarks`.

REQUIRED:
- [ ] All changes are up to date in `CHANGELOG.md`. **n/a**

Without vertex merging enabled:
![image](https://user-images.githubusercontent.com/1139788/115561831-c9c07b00-a26a-11eb-872f-6b86c5b62e1a.png)

With vertex merging enabled:
![image](https://user-images.githubusercontent.com/1139788/115561469-751d0000-a26a-11eb-9f84-bc1334c4fd10.png)


<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/hypar-io/elements/573)
<!-- Reviewable:end -->