@ikeough 2021-05-06
This is a problem. The vertex merging merges the UV coordinates. UV coordinate handling was not great to begin with due to the mistaken assumption that UV coordinates were blown away during CSG operations. They are not, which means that we can apply UV coordinates to the solid vertices, avoid tessellating in the case of the swept curve (the outer cylinder) and properly apply normals and UVs as the CSG is constructed. We might not need to do mesh re-merging at all.

![image](https://user-images.githubusercontent.com/1139788/117343790-80228380-ae59-11eb-8c62-f73f7c7c215b.png)